requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
blinker==1.6.2
google-generativeai